pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.2.0"
description = ""
authors = ["kate kate <ekaterina@chapkina.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0"
gunicorn = "^20.1.0"
dj-database-url = "^0.5.0"
python-dotenv = "^0.19.2"
pytest-django = "^4.5.2"
django-filter = "^23.1"
django-tables2 = "^2.4.1"
django-bootstrap4 = "^3.0.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
pytest = "^6.2.5"
pytest-cov = "^3.0.0"
coverage = "^7.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
